+++ src/Bcgov/EmergencyInfo/blocks/post-meta-display/index.js
@@ -1,0 +1,11 @@
+/**
+ * WordPress dependencies
+ */
+import { registerBlockType } from '@wordpress/blocks';
+import metadata from './block.json';
+import Edit from './edit';
+
+registerBlockType(metadata, {
+    edit: Edit,
+    save: () => null,
+});